@@ -4,17 +4,24 @@
   "description": "Demo Web App for learning purpose",
   "main": "index.js",
   "scripts": {
-    "test": "jest"
+    "build": "babel interest_calculator/src --out-dir dist --source-maps",
+    "watch": "babel interest_calculator/src --out-dir dist --source-maps --watch",
+    "test": "jest",
+    "lint": "eslint .",
+    "lint:fix": "eslint --fix .",
+    "pretty": "prettier --check .",
+    "pretty:fix": "prettier --write ."
   },
-  "type": "commonjs",
+  "type": "module",
   "keywords": [],
   "author": "",
   "license": "ISC",
   "devDependencies": {
+    "@babel/cli": "^7.26.4",
     "@babel/preset-env": "^7.26.0",
     "@eslint/js": "^9.17.0",
     "babel-jest": "^29.7.0",
-    "eslint": "^9.17.0",
+    "eslint": "^9.18.0",
     "eslint-plugin-react": "^7.37.3",
     "globals": "^15.14.0",
     "jest": "^29.7.0",
