@@ -1,3 +1,4 @@
 export default {
     presets: [["@babel/preset-env", { targets: { node: "current" } }]],
-};
+    sourceRoot: 'src', // Remove or adjust this line
+  };
