@@ -1,4 +1,4 @@
 module.exports = {
-    testEnvironment: "jest-environment-jsdom", // Ensure this is set
-  };
-  
+  testEnvironment: "jest-environment-jsdom",
+  moduleFileExtensions: ['js', 'jsx', 'json', 'node', 'mjs'], // Add 'mjs' for ESM support
+};
